created commands.test.ts
